% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{applyConsensus}
\alias{applyConsensus}
\title{apply the consensus to the segments}
\usage{
applyConsensus(segments, consensus)
}
\arguments{
\item{segments}{(cnvstrml)  the segments that will be further annotated}

\item{consensus}{(cnvstrml) with the consensus}
}
\value{
the segments with updated ranges and metadata to reflect if a segment is in consensus
}
\description{
this function splits each GRanges object in the segments
to the consensus segment and the unique part
The 'hit' column in the metadata is set to TRUE if it is found in the consensus
It is suggested to run this function \emph{before} filling the gaps with the function fillGaps()
}
