% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvstrml.R
\name{check_cnvstrml}
\alias{check_cnvstrml}
\title{check the validity of the inputs for the cnvstrml object}
\usage{
check_cnvstrml(object)
}
\arguments{
\item{object}{the cnvstrml object ot check}
}
\value{
TRUE or errors
}
\description{
check the validity of the inputs for the cnvstrml object
}
