% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvstrml.R
\docType{class}
\name{cnvstrml-class}
\alias{cnvstrml-class}
\alias{cnvstrml}
\title{create the a CNVstrml object}
\description{
cnvstrml is a structure that holds the information for the CNV predictions
and some additional information provided by the methods
}
\section{Slots}{

\describe{
\item{\code{segments}}{set of CNV segments (GRanges) as predicted by a CNV calling method}

\item{\code{purity}}{(real) the purity of the sample (default 1)}

\item{\code{ploidy}}{(real) the ploidy of the sample (default 2)}

\item{\code{genome}}{(character) the name of the genome to use}

\item{\code{sample}}{(character) name of sample}

\item{\code{method}}{(character) the name of the method used to produce the data}

\item{\code{sex}}{(character) the sex of the sample}

\item{\code{segtype}}{(character) the type of data stored in this object}
}}

\section{Notes}{

This is the main object provided by this package.

Contains information about the CNV segments
the purity and ploidy of the samples,
and metadata about the method name, sample name and the
reference genome used
}

