% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvstrmliner-package.R
\docType{package}
\name{cnvstrmliner-package}
\alias{cnvstrmliner}
\alias{cnvstrmliner-package}
\title{cnvstrmliner: What the Package Does (Title Case)}
\description{
CNV streamliner. The aim of this package is to streamline the analysis of results produced by CNV tools. It introduces the cnvstrml object and parsers for commonly used methods. It has some methods to compare different calls.
}
\keyword{internal}
