% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compareSegmentStatus}
\alias{compareSegmentStatus}
\title{compare the segments between two cnvstrml objects}
\usage{
compareSegmentStatus(query, subject, wiggle = 0.1)
}
\arguments{
\item{query}{(cnvstrml)}

\item{subject}{(cnvstrml)}

\item{wiggle}{(float) the difference between ccfs that are allowed to be considered the same (e.g. with wiggle =0.1 ccf 0.3 and 0.39 are considered the same)}
}
\value{
a table with numbers of TP, FP, FN
}
\description{
Compares the query to the subject segments
Finds teh common ranges, i.e. overlapping using findOverlaps
it performs the comparisons at three levels
\itemize{
\item cnv status (e.g. AMP, NEUT, DEL etc)
\item cnv copies (integer number of copies). This is applied only when status TP
\item cnv copies and ccf. CCF is adjusted by wiggle. This is applied only when status is TP
TP
}
}
