% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{fillGaps}
\alias{fillGaps}
\title{fill the gaps between the segments with segments of neutral copy number}
\usage{
fillGaps(segments, ploidy = 2, major.allele.copies = 1)
}
\arguments{
\item{segments}{(cnvstrml) the object which will be modified}

\item{ploidy}{(integer) expected copy number of the neutral region (default=2)}

\item{major.allele.copies}{(integer) expected number of the major allele in the neutral regions (default =1)}
}
\value{
an updated copy of the input object
}
\description{
This function finds the gaps between the segments that have been called
and adds them as new segments with neutral copy number.
It is suggested to run this function \emph{after} applying the consensus with the function applyConsensus()
}
