% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{findConsensus}
\alias{findConsensus}
\title{get the consensus sequence from a list of cnvstrml objects that represent the CNV segments}
\usage{
findConsensus(
  segmentsList,
  ignore.samples = FALSE,
  method = c("qualitative", "quantitative")
)
}
\arguments{
\item{segmentsList}{(list) a list of cnvstrml objects}

\item{ignore.samples}{(logical) if set to FALSE then the sample name of each element in segmentsList has to be the same (default FALSE)}

\item{method}{(character)}
}
\value{
a cnvstrml with all the segments that are shared across all samples.
}
\description{
The consensus is a new Granges that contains only the shared segments
The metadata of the consensus have the cnv.annotations for each of the elements of the segments
}
