% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{findConsensusGR}
\alias{findConsensusGR}
\title{get the consensus sequence from a list of GRanges objects that represent the CNV segments}
\usage{
findConsensusGR(segmentsList, method = c("qualitative", "quantitative"))
}
\arguments{
\item{segmentsList}{(list of GRanges) a list of cnvstrml objects}

\item{method}{(character) can be either 'qualitative' of 'quantitative'. If it is set to
'qualitative' it returns an object with qualitative information about the segments (e.g NEUT, AMP etc)
If it is set to 'quantitative' it retuns an object with the actual number of copy numbers.}
}
\value{
a Granges with all the segments that are shared across all samples.
}
\description{
The consensus is a new Granges that contains only the shared segments
The metadata of the consensus have the cnv.annotations for each of the elements of the segments
}
