% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\docType{methods}
\name{getPloidy}
\alias{getPloidy}
\title{Get the  ploidy for the sample}
\usage{
getPloidy(.Object, rounded)

\S4method{getPloidy}{cnvstrml,logical}(.Object, rounded)

\S4method{getPloidy}{cnvstrml,ANY}(.Object, rounded)
}
\arguments{
\item{.Object}{object of type cnvstrml,}

\item{rounded}{get the rounded ploidy (default TRUE)}
}
\value{
the ploidy
}
\description{
Two parameters can be retrieved:
}
\details{
\strong{rounded ploidy} is the rounded ploidy (e.g. 2.1 will become 2). This is the default.

\strong{ploidy} is the actual value provided by the user
}
