% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity.R
\docType{methods}
\name{getPurity}
\alias{getPurity}
\alias{getPurity,cnvstrml-method}
\title{Get the  purity for the sample}
\usage{
getPurity(.Object)

\S4method{getPurity}{cnvstrml}(.Object)
}
\arguments{
\item{.Object}{object of type cnvstrml,}
}
\value{
the purity of the sample (0 - 1)
}
\description{
Two parameters can be retrieved:
}
\details{
\strong{purity} is the actual value provided by the user
}
