% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\docType{methods}
\name{getSex}
\alias{getSex}
\alias{getSex,cnvstrml-method}
\title{Get the  sex for the sample}
\usage{
getSex(.Object)

\S4method{getSex}{cnvstrml}(.Object)
}
\arguments{
\item{.Object}{object of type cnvstrml,}
}
\value{
the sex (M for male, F for female)
}
\description{
Get the  sex for the sample

Get the  sex for the sample
}
