% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\docType{methods}
\name{inferSex}
\alias{inferSex}
\alias{inferSex,cnvstrml-method}
\title{Infer the sex for a sample}
\usage{
inferSex(.Object)

\S4method{inferSex}{cnvstrml}(.Object)
}
\arguments{
\item{.Object}{object of type GRanges or GRangeslist}
}
\value{
the sex of the sample as XX for female and XY for male
}
\description{
If events occur on the Y chromosome the sample
is assigned to XY (male)
otherwise XX (female)
}
\section{Notes}{

The inference of the sex depends on the identification of events on the sex chromosome
If any event is observed on chrY then it is assumed that the sample is male
The return value is M for male or F for female
}

