% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvstrml.R
\name{initialize,cnvstrml-method}
\alias{initialize,cnvstrml-method}
\title{initialize the object}
\usage{
\S4method{initialize}{cnvstrml}(.Object, ...)
}
\arguments{
\item{.Object}{an .Objectect of class cnvstrml}

\item{...}{additional arguments}
}
\value{
an initialized .Objectect
}
\description{
check that purity is within limits
updates the sex based on the presence of chrY events if the user has not prodived that
}
