% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseManta.R
\name{matchCNVSV}
\alias{matchCNVSV}
\title{In order to match the SV breakpoints to the CNV segments
we expect to find a breakpoint in the vicinity of the
starting and ending coordinates of the CNV segment
Furthermore the two breakpoints need to be partners.}
\usage{
matchCNVSV(segs, sv, wiggle = 1000)
}
\arguments{
\item{sv}{(cnvstrml) the SV ranges from breakpointRanges}

\item{wiggle}{(numeric) distance between the CNV boundaries and sv breakpoints}

\item{segments}{(cnvstrml) the segments predicte to be copy altered}
}
\value{
an updates GRanges with segments and support from SV
}
\description{
In order to match the SV breakpoints to the CNV segments
we expect to find a breakpoint in the vicinity of the
starting and ending coordinates of the CNV segment
Furthermore the two breakpoints need to be partners.
}
