% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FREEC.R
\name{parseCfreec}
\alias{parseCfreec}
\title{Generate a cnv Granges from a dataframe created by ControlFreec (file ending in .CNVs or contains pvalues.)
Control-FREEC is a tool for detection of copy-number changes and allelic imbalances (including LOH)
using deep-sequencing data. It automatically computes, normalizes, segments copy number and beta allele
frequency (BAF) profiles, then calls copy number alterations and LOH. The control (matched normal) sample
is optional for whole genome sequencing data but mandatory for whole exome or targeted sequencing data.
Control-FREEC up to now (Jan 2021) is the standard tool used in the hCELG and hBMS pipelines for data
analysis of WES and WGS datasets.
When the genotype is - it means that the progrm did not find any SNPs to decide the correct genotype, and
the estimated copy number depends on the logRR
The output includes
\itemize{
\item chromosome
\item start position
\item end position
\item copy.number
\item status
\item genotype
\item uncertainty of the assignment
\item WilcoxonRankSumTestPvalue
\item KolmogorovSmirnovPvalue
Note: at least for v 11.5 it is not possible to understand how subclones are coded in the output files.
}}
\usage{
parseCfreec(controlfreec.df, filterLowUncertainty = TRUE, pvalThreshold = 0.01)
}
\arguments{
\item{controlfreec.df}{input filename with CNV segments,}

\item{pvalThreshold}{threshold of pvalues to filter if the pvalues are included in the file}

\item{filterLowUncerntainty}{if set to TRUE segments with uncertainty equal to -1 will be discarded}
}
\value{
A GRanges object with the segments predicted by ControlFreec
}
\description{
Generate a cnv Granges from a dataframe created by ControlFreec (file ending in .CNVs or contains pvalues.)
Control-FREEC is a tool for detection of copy-number changes and allelic imbalances (including LOH)
using deep-sequencing data. It automatically computes, normalizes, segments copy number and beta allele
frequency (BAF) profiles, then calls copy number alterations and LOH. The control (matched normal) sample
is optional for whole genome sequencing data but mandatory for whole exome or targeted sequencing data.
Control-FREEC up to now (Jan 2021) is the standard tool used in the hCELG and hBMS pipelines for data
analysis of WES and WGS datasets.
When the genotype is - it means that the progrm did not find any SNPs to decide the correct genotype, and
the estimated copy number depends on the logRR
The output includes
\itemize{
\item chromosome
\item start position
\item end position
\item copy.number
\item status
\item genotype
\item uncertainty of the assignment
\item WilcoxonRankSumTestPvalue
\item KolmogorovSmirnovPvalue
Note: at least for v 11.5 it is not possible to understand how subclones are coded in the output files.
}
}
