% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACETS.R
\name{parseFACETS}
\alias{parseFACETS}
\title{parse the output of FACETS}
\usage{
parseFACETS(facets.data, sample = NA_character_, genome = "hg38")
}
\arguments{
\item{facets.data}{(character or data.frame) the input data.
if the type of the argument is character it is assumed to be a file.
In this case it expects the extension .vcf or .vcf.gz to parse it as a vcf file.
if the type of the argument is data.frame it is parsed as such.}

\item{sample}{(sample)  the name of the sample. If the input is a vcf file it will use the name of the sample in the file.}

\item{genome}{(character)  the genome version assumed for this dataset}
}
\value{
a cnvstrml object
}
\description{
this function is a high level parser that can handle
either data frames from FACETS or vcf files
}
