% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACETS.R
\name{parseFACETS_df}
\alias{parseFACETS_df}
\title{create a cnvstrml object from a dataframe with facets data}
\usage{
parseFACETS_df(
  df,
  purity = NA_real_,
  ploidy = NA_real_,
  sample = NA_character_,
  genome = "hg38"
)
}
\arguments{
\item{df}{(data.frame)  dataframe produced by FACETS}

\item{purity}{(numeric)  the purity produces by FACETS (a value between 0 - 1)}

\item{ploidy}{(numeric) the ploidy produced by FACETS}

\item{sample}{(character)  the name of the sample}

\item{genome}{(character) the genome version assumed for this dataset}
}
\value{
a cnvstrml object
}
\description{
create a cnvstrml object from a dataframe with facets data
}
