% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FACETS.R
\name{parseFACETS_vcf}
\alias{parseFACETS_vcf}
\title{parse a VCF file in BND format with FACETS data}
\usage{
parseFACETS_vcf(vcf, sample = NA_character_, genome = "hg38")
}
\arguments{
\item{vcf}{(character) the input VCF file}

\item{sample}{(character)  the name of the sample}

\item{genome}{(character)  the genome version assumed for this dataset}
}
\value{
a cnvstrml object
}
\description{
Such a file is created from the cnv_facets tool
}
