% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseManta.R
\name{parseManta}
\alias{parseManta}
\title{Parse the manta ouput (vcf v4.1)}
\usage{
parseManta(sv.vcf, sample, genome = "hg38")
}
\arguments{
\item{sv.vcf}{(character) a file name (in vcf format)}

\item{sample}{(character) the name of the sample}

\item{genome}{(character) the genome version to use (hg38). This is passed to VariantAnnotation::readVcf}
}
\value{
a cnvstrml object, Purity and Ploidy are not set
}
\description{
Parse the manta ouput (vcf v4.1)
}
