% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSegments}
\alias{plotSegments}
\title{plot the chromosomes with the segments with copy number}
\usage{
plotSegments(segments, max.copies = 6, main = NULL)
}
\arguments{
\item{segments}{(cnvstrml) the object to plot}

\item{max.copies}{(integer) the maximum number of copies to plot. Any segment with more copies will be downsized to this number. The Y scale of the plot ranges between 0 and max.copies.}
}
\description{
plot the chromosomes with the segments with copy number
}
