% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvstrml.R
\name{show,cnvstrml-method}
\alias{show,cnvstrml-method}
\title{show the contents of the object}
\usage{
\S4method{show}{cnvstrml}(object)
}
\arguments{
\item{object}{Object of type cnvstrml}
}
\description{
show the contents of the object
}
